게시판 완성

+++ src/main/java/ks46team01/user/board/service/FreeBoardService.java
@@ -18,6 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -78,8 +79,20 @@ public class FreeBoardService {
     }
 
 
+
+    @Transactional
     public boolean deleteFreeBoard(Long freeBoardIdx) {
         try {
+
+            FreeBoard freeBoard = freeBoardRepository.findById(freeBoardIdx).orElse(null);
+            if (freeBoard == null) {
+                throw new IllegalArgumentException("Invalid freeBoardIdx");
+            }
+
+
+            freeBoardReplyRepository.deleteAllByFreeBoard(freeBoard);
+
+            // Now delete the FreeBoard
             freeBoardRepository.deleteById(freeBoardIdx);
             return true;
         } catch (Exception e) {
